Created Promotions-table DB migration
Nothing too fancy-schmancy here. Modelling a promotion means (1)
the moola threshold you'd like the promotion in question to begin
kicking in above, and (2) the percentage you'd like the component of
the shopping cart's total above that threshold reduced by.

We're assuming that this is the only kind of promotion we'd like so
far. Just a simple percentage? It's not that difficult to implement other
types: fixed-amount discounts, etc. But the spec calls, at this point,
only for percentages and thresholds. Let's go with those.

Also, thinking ahead to UI: what kind of button or twiddly-bit(s)
might we need? Standard promotions/discounts often have something
like "Enter a discount code here to receive Promotion X, baby". Do
we need a code? Spec says zilch. Let us kick that can down our
codey road.