Populated seeds.db
A few things going on here. The Seeds file is a finicky
beastie, isn't it. The whole point is to populate the dev
database with *something* useful and play-about-able,
though not necessarily actual real data.

One thing that'd always irked me is side effects from
seeding the database multiple times. My first rule of
thumb is to first empty the contents of any table you're
seeding. Cleaner.

Wait, scratch that. My actual first rule is to make absolutely
and inviolably sure that you're not attempting to seed a
database in a setting or an environment that could
possibly hold valuable data. Staging/Prod. Before I run
anything at all, check the environment, and if we're outside
Development/Testing, taunt the user. Without remorse
or pity.

Okay, done. Now add equivalent contents of our json
file and Promotions spec. It's okay, we'll add the file upload
UI fiddly-bit as per the spec soon enough. This seed is
just useful demo data.