Created CartItem DB-table migration
Had a bit more of a think about what we're building here. It's
a single shopping cart, operated by a single user. The user
can select from a pool of existing products, add them to a
shopping cart, select the number of each product they are
adding (not part of the spec in all fairness but what
self-respecting cart is without that), apply promotions (manual
or automatic is not in the spec, but shopping carts typically
have a "Enter Code XXX to Receive Promotion" field, so
I'll build that, and assume the promotion is manually triggered.

The UI fiddly bits can come later. Let's implement the
CartItems. Each has: product ID foreign key; quantity of that
product.