Refactor Rask task file into unified "prepare app" file
I wouldn't normally do this in a professional app, but I
figure a decent wireframe is to have just a single
ShoppingCart object, and use its attributes to store
our app's data. If we were to ever extend this in a proper
real-world app, we'd naturally wish to have multiple
shopping carts, likely one per user, so this has at least
some extensibility.