CartItem CRUD functionality added
Bit more substantial! I had to get through a fair bit
of Rails Guides trawling to remember how to code
this stuff. But that was my plan all along, so mission
accomplished.

Bit of explanation in the routes: I could have added
`:increment` and `:decrement` to the `only:` array,
but this would not have created the named routes I've
then used in the +/-/x single-button forms for each
cart item. Routes like
`shopping_cart_cart_item_increment_path(...)` are
a bit of a mouthful, but better than manually
hardcoding each form URL.