Implemented ShoppingCart#total and submethods
This took quite a lot of headbutting! There is absolutely
something to be said for well-written, testable, tested
code. But adding actual unit test suites on top of
everything else would shove the work hours required
for this MVP way, way above the "just a few hours"
you'd asked of me. This is a hello-world MVP.