Sidestepping apparent bug in either Rails controller or view
Not sure which! shopping_cart_cart_item_increment_path
and shopping_cart_cart_item_path both receive a cart_item
object, but when writing code inside the controller methods
they point to, the exact same cart_item object populates
`params[:cart_item_id]` on the former, but `params[:cart_id]`
on the latter. Could be because `increment` is a custom,
non-CRUD route? Search me. I don't doubt we could figure
out how to manually populate each `params` key name in
the bowels of either Rails's controller or routes docs, but
I think this is beyond the scope of this MVP. This hack will
do for now.